catch additional warnings
