+++ tests/__init__.py
@@ -1,8 +1,10 @@
+import re
 from dataclasses import dataclass, field
 from enum import Enum
 from typing import Any, Dict, List, Optional, Union
 
 import attr
+from pytest import warns
 
 from omegaconf import II, MISSING
 
@@ -220,3 +222,12 @@ class InterpolationDict:
 @dataclass
 class Str2Int(Dict[str, int]):
     pass
+
+
+def warns_dict_subclass_deprecated() -> Any:
+    return warns(
+        UserWarning,
+        match=re.escape(
+            "Subclassing of `Dict` by Structured Config classes is deprecated"
+        ),
+    )
